requirements.txt
volcasample[audio]

Flask
Flask-SQLAlchemy
Flask-WTF
sqlalchemy-migrate
python-dotenv
simpleaudio
celery
urllib3
redis
future